### 2021-08-02 6191323
Update lammps_data.py
removing mpi as it creates complications installing on HPCs
