+++ requirements.txt
@@ -2,7 +2,6 @@ pymatgen>=v2017.10.16
 numpy
 matplotlib>=3.1.0
 scipy
-mpi4py
 seaborn>=0.9.0
 pytest
 mock
