### 2021-08-03 54554ec
updating readme instructions for conda installing on Archer2
