+++ README.md
@@ -6,12 +6,11 @@
 ## Overview
 PopOff currently fits Coulomb-Buckingham type interatomic potentials for classical potential-based molecular dynamics (MD). This is a modular fitting code, allowing increased control over several important aspects of the potenial. You can choose between fitting a formal charge, partial charge, or fitting a charge scaling factor. You can choose to fit a rigid ion model or a core-shell model. If choosing to fit a core-shell model, you are able to fix or fit the charge separation over the core and shell components and the spring constant. You can also selectively fix, fit, or set to zero the individual buckingham parameters.
 
-## Prelude
+## Prerequisites
+To run the potential fitting code, you need LAMMPS with the LAMMPS-python interface package. LAMMPS is relatively easy to install and so is the python interface if you are working locally. The python interface can be challenging to compile on HPC or larger servers and you may need assistance from the dedicated HPC support team. Here, I provide the basic instructions for compiling LAMMPS and the potential fitting code locally. I also provide instructions on how to download and install the code in a conda enviroment to run on Archer2 (which may work for other HPC systems). 
 
-To run the potential fitting code, you must first download LAMMPS and install at a minimum the LAMMPS python package. Installing LAMMPS itself is a relatively easy task. You just download the relevant source files, go to the src directory, and run make. Simple.
-
-## Compiling LAMMPS
-Download the version of LAMMPS you want, making sure it is a release from March 23rd 2018. Before this date some of the libraries weren't properly supported. You can either `git clone` the latest stable version using:
+## Compiling LAMMPS and the code locally
+Download the version of LAMMPS you want, making sure it is a release from March 23rd 2018 onwards. Before this date some of the libraries weren't properly supported. You can either `git clone` the latest stable version using:
 
 ```
 git clone -b stable https://github.com/lammps/lammps.git [path to build location]/lammps
@@ -26,23 +25,23 @@ export LAMMPS_PACKAGES="asphere body class2 colloid compress coreshell dipole gr
 for pack in $LAMMPS_PACKAGES; do make "yes-$pack"; done
 ```
 
-These are the packages installed during development and testing of this code, though you should note that some packages are not compatible with others and can cause conflicts. So just be careful.
+These are the packages installed during development and testing of this code, though you should note that some packages are not compatible with others and can cause conflicts. So, just be careful.
 
 Right, now you are ready to install LAMMPS!
 
-This is done using a `make` command. Here you can also chose which mode to use, to run as serial or mpi, and which flags to use. There are suggested flags to use to get everything working smoothly. The make command I would suggest using is:
+This is done using a `make` command (you can also use cmake if preferred). Here you can also chose which mode to use, to run as serial or mpi, and which flags to use. There are suggested flags to use to get everything working smoothly. The make command I would suggest using is:
 
 ```
 make mode=shlib mpi -j4 LMP_INC="-DLAMMPS_EXCEPTIONS -DLAMMPS_GZIP -DLAMMPS_MEMALIGN=64 -DBUILD_SHARED_LIBS=on"
 ```
+
+Boom! Now you are compiling LAMMPS. Make a cup of tea (or coffee if you prefer) while it runs. This could take about 5 minutes, or longer if you have more packages.
 
 To install the python interface package, you can do this using:
 
 ```
 make install-python
 ```
-
-Boom! Now you are compiling LAMMPS. Make a cup of tea (or coffee if you prefer) while it runs. This could take about 5 minutes, or longer if you have more packages.
 
[…]
