### 2021-08-05 #3451 @dpsanders
👋 @LMMorgan: Apologies for the delay in getting to this. This looks like a nice package and paper!

Here are some initial comments:

- I suggest to reorder the text so that a non-specialist reader immediately understands the context, e.g. by putting the Statement of Need first

- Please explain what VASP and LAMMPS are, with references, and what they output that you then fit.

- Do you have references to the research applications?

- differentail -> differential (line 68)

- I don’t see where the docs are for the package. Maybe this is just a missing link from the GitHub repo, since there is a `docs` folder?
