### 2021-08-05 #3451 @dpsanders
Also I personally like to see a small working example of the package in the paper, with example output.
