### 2021-08-09 #3451 @LMMorgan
Hi @dpsanders thank you for getting back to me. I understand you are all quite busy after the hiatus and things are likely to be fairly slow.

In regards to your comments. I shall start addressing your points this week. I am not sure where the docs link has disappeared to.

A quick query, when you say "a small working example of the package in the paper, with example output", do you mean you would like a working example directory in the repo? (I should be able to do this fairly swiftly for Na3OCl or another structure) I have not seen anyone present a working example written into the paper before, but if the latter is what you are asking for, could you send an example where this has been done, so to use as a guide.

Many thanks,
Lucy
