### 2021-08-11 18ac653
updating paper per the editors request. Adding citations for VASP and LAMMPS, checking PopOff/popoff consistency, correcting typo, and updating the installation instructions in rst
