+++ paper/paper.md
@@ -33,20 +33,20 @@ date: 2 July 2021
 bibliography: paper.bib
 ---
 
-# Summary
-`PopOff` is a Python package for fitting Coulomb-Buckingham type interatomic potentials for classical potential-based molecular dynamics (MD). This is a modular fitting code, allowing increased control over several important aspects of the potential. Formal charge, partial charge, or a charge scaling factor can be fitted. A rigid ion model, a core-shell model, or mixture of both can be chosen. If choosing to fit a core-shell model, the charge separation over the core and shell components can be fixed or fitted, as can the spring constant. Individual Buckingham parameters can also be selectively fixed, fittted, or set to zero. `PopOff` currently fits to first principles derived forces and stress tensors obtained from a VASP training set through the `vasprun.xml` output files. The LAMMPS MD code is used as a backend to perform the MD during the fitting process.
-
-The development of sufficiently accurate interatomic potentials, which requires model parameterization with respect to a set of target observables, for a specific chemistry is quite challenging, especially for complex systems. For example, in layered structures, such as the widely used LiNiO$_2$ based cathode materials, the short-range interactions are overwhelmed by the longer-range Coulombic term. In these cases, the system charges need to be scaled down to increase the influence of the short-range interactions, termed as partial charges, and the influence of polarizability needs to be included, usually through using a core-shell model. Variations of the Buckingham potential have been developed for these systems, some using rigid ion models,[@Lewis_1985; @Ledwaba2020; @Sayle2005; @Dawson2014] and others using core-shell models, [@Hart1998; @Fisher2010; @Lewis_1985; @Ammundsen1999; @Kerisit2014; @He2019; @lee2012atomistic] with a mixture of formal and partial charges being used. Interatomic potentials are traditionally based on mathematical functions which have been parameterized using experimental and/or First Principles derived data [@jones_1924; @buckingham_classical_1938]. There are a limited number of codes available with the explicit purpose or functionality for fitting potentials. Fitting routines from established codes including the General Utility Lattice Program (GULP), [@gale_gulp_1997] Atomicrex, [@Stukowski_2017] dftfit, [@dftfit] and potfit [@wen_kim-compliant_2017] each possess unique functionality, however, none are fully able to produce robust potentials for NMC or LiNiO$_2$ due to different aspects not being considered within the code, such as charge scaling or core-shell models. `PopOff` has been specifically developed for fitting different permutations of the Buckingham potential. Its modular design allows flexible fitting of both rigid ion and core-shell models, and formal and partial charges. 
-
-We are currently using `PopOff` in our own research to develop potentials for cathode and solid-electrolyte materials for classical MD analysis of Li$^+$/Na$^+$ migration properties. `PopOff` is also being used by a group at Newcastle University to develop potentials for several halide based solid-electrolyte materials in collaboration with experimental partners. We hope that this open-source resource will support development of interatomic potentials tailored to individual systems, and thus results in better approximations of the atomic interactions.
-
 # Statement of need
[…]
