+++ paper/paper.bib
@@ -163,42 +163,42 @@
 }
 
 @article{Stukowski_2017,
-	doi = {10.1088/1361-651x/aa6ecf},
-	year = 2017,
-	month = {may},
-	publisher = {{IOP} Publishing},
-	volume = {25},
-	number = {5},
-	pages = {055003},
-	author = {Stukowski, Alexander and Fransson, Erik and Mock, Markus and Erhart, Paul},
-	title = {{Atomicrex--a general purpose tool for the construction of atomic interaction models}},
-	journal = {Modelling and Simulation in Materials Science and Engineering},
-	url = {https://www.atomicrex.org}
+    doi = {10.1088/1361-651x/aa6ecf},
+    year = 2017,
+    month = {may},
+    publisher = {{IOP} Publishing},
+    volume = {25},
+    number = {5},
+    pages = {055003},
+    author = {Stukowski, Alexander and Fransson, Erik and Mock, Markus and Erhart, Paul},
+    title = {{Atomicrex--a general purpose tool for the construction of atomic interaction models}},
+    journal = {Modelling and Simulation in Materials Science and Engineering},
+    url = {https://www.atomicrex.org}
 }
 
 @article{wen_kim-compliant_2017,
-	title = {{A KIM-compliant \textit{potfit} for fitting sloppy interatomic potentials: application to the EDIP model for silicon}},
-	volume = {25},
-	issn = {0965-0393, 1361-651X},
-	shorttitle = {A KIM-compliant \textit{potfit} for fitting sloppy interatomic potentials},
-	doi = {10.1088/0965-0393/25/1/014001},
-	number = {1},
-	journal = {Modelling and Simulation in Materials Science and Engineering},
-	author = {Wen, Mingjian and Li, Junhao and Brommer, Peter and Elliott, Ryan S. and Sethna, James P. and Tadmor, Ellad B.},
-	month = jan,
-	year = {2017},
-	pages = {014001}
+    title = {{A KIM-compliant \textit{potfit} for fitting sloppy interatomic potentials: application to the EDIP model for silicon}},
+    volume = {25},
+    issn = {0965-0393, 1361-651X},
+    shorttitle = {A KIM-compliant \textit{potfit} for fitting sloppy interatomic potentials},
+    doi = {10.1088/0965-0393/25/1/014001},
+    number = {1},
+    journal = {Modelling and Simulation in Materials Science and Engineering},
+    author = {Wen, Mingjian and Li, Junhao and Brommer, Peter and Elliott, Ryan S. and Sethna, James P. and Tadmor, Ellad B.},
+    month = jan,
+    year = {2017},
+    pages = {014001}
 }
 
 @article{jackson1985trapping,
-  title={{Trapping and solution of fission Xe in UO$_2$.: Part 1. Single gas atoms and solution from underpressurized bubbles}},
-  author={Jackson, R. A. and Catlow, C. R. A.},
-  journal={Journal of nuclear materials},
-  volume={127},
-  number={2-3},
-  pages={161--166},
-  year={1985},
-  publisher={Elsevier}
+    title={{Trapping and solution of fission Xe in UO$_2$.: Part 1. Single gas atoms and solution from underpressurized bubbles}},
+    author={Jackson, R. A. and Catlow, C. R. A.},
+    journal={Journal of nuclear materials},
+    volume={127},
+    number={2-3},
+    pages={161--166},
+    year={1985},
+    publisher={Elsevier}
 }
 
 @misc{dftfit,
@@ -206,4 +206,59 @@
     title = {{dftfit}},
     url = {https://chrisostrouchov.com/dftfit},
     urldate = {2020-09-29}
+}
+
+@article{kresse1993ab,
+    title={Ab initio molecular dynamics for liquid metals},
[…]
