### 2021-08-23 1660885
adding Na3OCl example and updating paper and docs
