+++ paper/paper.md
@@ -39,7 +39,7 @@ Interatomic potentials give an approximate mathematical description of the inter
 There are a limited number of codes available with the explicit purpose or functionality for fitting potentials. Most of which require some pre-requisite knowledge of coding or specialism, creating a prohibitive barrier for utilising these tools in the wider atomistic modelling research community. These codes are also not able to account for all aspects which may need to be considered for particular materials, such as polarizability. Creating a resource, with minimal pre-requisite requirements to use, and able to account for a larger range of material considerations needed for accurate potentials is therefore key to producing more robust classical MD investigations.
 
 # Summary
-`PopOff` is a Python package for fitting Coulomb-Buckingham type interatomic potentials for classical potential-based molecular dynamics (MD). This is a modular fitting code, allowing increased control over several important aspects of the potential. Formal charge, partial charge, or a charge scaling factor can be fitted. A rigid ion model, a core-shell model, or mixture of both can be chosen. If choosing to fit a core-shell model, the charge separation over the core and shell components can be fixed or fitted, as can the spring constant. Individual Buckingham parameters can also be selectively fixed, fittted, or set to zero. `PopOff` currently fits to first principles derived forces and stress tensors obtained from a VASP [@kresse1993ab; @kresse1994ab; @kresse1996efficiency; @kresse1996efficient] training set through the `vasprun.xml` output files. The LAMMPS MD code [@plimpton1995fast] is used as a backend to perform the MD during the fitting process and outputs the forces and stress tensors, produced using a range of potentials from the search space. The output atom forces and system stress tensors are compared and minimised.
+`PopOff` is a Python package for fitting Coulomb-Buckingham type interatomic potentials for classical potential-based molecular dynamics (MD). This is a modular fitting code, allowing increased control over several important aspects of the potential. Formal charge, partial charge, or a charge scaling factor can be fitted. A rigid ion model, a core-shell model, or mixture of both can be chosen. If choosing to fit a core-shell model, the charge separation over the core and shell components can be fixed or fitted, as can the spring constant. Individual Buckingham parameters can also be selectively fixed, fittted, or set to zero. `PopOff` currently fits to first principles derived forces and stress tensors obtained from a VASP (The Vienna Ab initio Simulation Package: atomic scale materials modelling from first principles) [@kresse1993ab; @kresse1994ab; @kresse1996efficiency; @kresse1996efficient] training set through the `vasprun.xml` output files. LAMMPS (Large-scale Atomic/Molecular Massively Parallel Simulator), [@plimpton1995fast] a classical MD code with a focus on materials modeling, is used as a backend to perform the MD during the fitting process and outputs the forces and stress tensors, produced using a range of potentials from the search space. The output atom forces and system stress tensors are compared and minimised.
 
[…]
