+++ userguides/example_Na3OCl/output/potential_fitting.py
@@ -1,0 +1,99 @@
+#! /usr/bin/env python3
+
+from buckfit.fitting_code import FitModel
+import buckfit.fitting_output as output
+from buckfit.input_checker import setup_error_checks
+from scipy import optimize
+import numpy as np
+
+def random_set_of_structures(fits, structures, structures_to_fit, seed=False):
+    """
+    Randomly selects structures up to the number of structures to include in the fit, checks there are no repeats and returns the structure numbers of those to be included in the fit.
+    Args:
+        fits (int): Number of fits to run.
+        structures (int): Total number of structures in the training set.
+        structures_to_fit (int): Number of structures to fit to.
+        seed (optional: int or 1-d array_like): Seed for RandomState. Must be convertible to 32 bit unsigned integers. Default=False
+    Returns:
+        (np.array): Structure numbers of structures in training set to be fitted to.
+    """
+ 
+    for x in [fits, structures, structures_to_fit]:
+        if isinstance(x, int) == False:
+            raise TypeError('Input values must be integers.')
+        if x <=0:
+            raise ValueError('Integer values must be positive.')
+
+    if seed is not False:
+        if isinstance(seed, int) == False and isinstance(seed, np.ndarray) == False:
+            raise TypeError('seed must be an integer, 1-d array, or False')
+        else:
+            np.random.seed(seed)
+            sets_of_structures = []
+            while len(sets_of_structures) < fits:
+                struct_set = np.sort(np.random.randint(0, structures, size=structures_to_fit), axis=0)
+                if len(set(struct_set)) != structures_to_fit:
+                    continue
+                if not any(np.array_equiv(struct_set, x) for x in sets_of_structures):
+                    sets_of_structures.append(struct_set) 
+            return np.array(sets_of_structures)
+
+def run_fit(sets_of_structures, params, labels, bounds, supercell=None, seed=None):
+    """
+    Collates the structures to be fitted into the working directory, creates the lammps inputs and runs the optimiser to fit to the designated parameters. Calls another function to save the data in the appropriate output directory.
+    Args:
+        sets_of_structures (np.array): Structure numbers of structures in training set to be fitted to.
+        params (dict(dict)): Setup dictionary containing the inputs for coreshell, charges, masses, potentials, and core-shell springs.
+        labels (list(str)): List of parameters to be fitted.
+        bounds (list(tuple(float))): List of lower and upper bound tuples associated with each parameter.
+        supercell (optional:list(int) or list(list(int))): 3 integers defining the cell increase in x, y, and z for all structures, or a list of lists where each list is 3 integers defining the cell increase in x, y, z, for each individual structure in the fitting process. i.e. all increase by the same amount, or each structure increased but different amounts. Default=None.
+        seed (optional: int or np.random.RandomState): If seed is not specified the np.RandomState singleton is used. If seed is an int, a new np.random.RandomState instance is used, seeded with seed. If seed is already a np.random.RandomState instance, then that np.random.RandomState instance is used. Specify seed for repeatable minimizations.
+    Returns:
+        None
+    """
+    
+    for fit, structs in enumerate(sets_of_structures):
[…]
