### 2021-08-23 #3451 @LMMorgan
Dear @dpsanders,

I have reordered the text and explained VASP and LAMMPS, with references. I have also added the missing `docs` link in the repo (it got missed off the previous commit) and I have adding the input and output files for an example (Na3OCl) to `userguides/example_Na3OCl`.

In regards to `PopOff` vs `popoff`, the lower case is used in code snippets as this in how the directories are labelled, and the capitalised is use in text as the formal acronym for the code. I think there was one place where this was incorrect and I have now corrected it.

I have also added a guide to compiling the cod and prerequisites in a conda environment for use on HPCs, as it can be challenging to compile LAMMPS with a python interface if you don't know what you are doing.

Please let me know if there are any other updates you would like.

Kind regards,
Lucy
