### 2021-09-20 #3451 @dpsanders
👋 @LMMorgan: Apologies for the delay in getting to this.

The paper is much improved, thanks! I think that`popoff` in the title itself should be `PopOff`?

I now see the docs link in the repo, thanks.
I see that the example notebooks require a VASP license and an HPC.
Is it possible to run a simpler example using only open-source software? This is not a requirement for acceptance in JOSS, but would definitely make the paper more broadly useful.

Thanks!
