### 2021-09-21 #3747 @whedon
[REVIEW]: popoff: POtential Parameter Optimisation for Force-Fields
**Submitting author:** @LMMorgan (<a href="http://orcid.org/0000-0002-6432-3760">Lucy M. Morgan</a>)
**Repository:** https://github.com/LMMorgan/PopOff
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
