### 2021-09-25 #3451 @LMMorgan
> I see that the example notebooks require a VASP license and an HPC.
> Is it possible to run a simpler example using only open-source software? This is not a requirement for acceptance in JOSS, but would definitely make the paper more broadly useful.

Hi @dpsanders, thank you for the update. The example I have given includes the VASP input files so there is no requirement to have a licence to run the example. If a user wanted to run their own potential fit from their dataset, they would need a vast licence or convert their output to a `vasprun.xml` format. I am hoping to generalise the input to include open-source in the next version of the code.

@marshallmcdonnell and @lscalfi, thank you for agreeing to be reviewers. I look forward to hearing from you.

Kind regards,
Lucy
