### 2022-01-16 #3747 @lscalfi
Dear @LMMorgan, sorry for the delay, I had a hard time installing the different packages.

The PopOff library is a python library for fitting force field parameters for Buckingham potentials, possibly using core-shell models. It is interfaced with the molecular dynamic code LAMMPS using the LAMMPS python interface. This adds a (complex) dependency to the PopOff library but frees the user from the need to build an interface to an MD code. The fitting is done with respect to DFT forces and stress tensors, obtained from VASP DFT simulations. This library is thought for application to bulk materials but could be easily applied to other systems. These are my remarks and questions on the paper, documentation and the library itself:

- One of the main criticisms I have is that the functionality of the code is not enough explained and should be more detailed both in the paper and in the docs/readme. In particular, although known, the potential form of the Coulomb-Buckingham potential should be written down, and each fitting parameter clearly identified (called a, rho and c in the notebooks).
Similarly, the core-shell model should be introduced more in details. The links to the corresponding LAMMPS potentials could be useful too.
Moreover, the informations are somehow slitted between the paper, the various readme, the documentation and the Jupyter notebooks, which makes it difficult to quickly grasp all the functionalities of the code, for example what parameters can be fitted, and what are the 'different permutations of the Buckingham potential' considered.
Also the details on the optimizer are only given in the paper but could be added or centralized in the docs.

- The paper mentions other software to fit interatomic potentials, could more details on their functionalities be included?
Also, why only Buckingham potentials can be fitted? It seems it would be easy to generalize this to other potentials implemented in LAMMPS.

- The readme file should be completed: it only contains instructions to build the LAMMPS python library. It should also give instructions to build the PopOff package and it would be useful to have a few lines on how to find and build the docs. In the docs, it would be easier if the installation instructions for the LAMMPS python library and PopOff were given together (instead of in the Dependencies).

- Which LAMMPS packages are needed for PopOff and for testing purposes? The packages 'snap' and 'user-reaxc' couldn't be installed in my LAMMPS (Package does not exist error) but this didn't cause problems for running the tests.

- The documentation and readme uses 'I', this should probably be updated.

- There is a section for asking help at the end of the docs but no community guidelines for developers (merge requests allowed?...). These informations should be added.

- Could you cite works where this package has been used? Please also specify the mention to 'a group at Newcastle University'.

- The sentence in the paper 'This can be changed in fitting_code.py, however, this is not given as a suggested change in the manual.' is confusing. Should these changes be suggested? Are they hard-coded in the package?

- Regarding the Jupyter notebooks, I was unable to fully run them. Indeed, the example given is Na3OCl, but the notebooks are written for LiNiO2. I tried running the script 'userguides/example_Na3OCl/output/potential_fitting.py' but it is apparently an old version. Indeed, it calls the library BuckFit (I guess an old version that should maybe be mentioned somewhere). When changed to PopOff, I have a weird FileNotFoundError in lammps_data.py, line 252, when trying to open the 'lammps/coords1.lmp' file for writing.
Please provide either the files for LiNiO2 or change the notebooks to fit to the NaO3Cl example.
Nevertheless, I already have some questions:
- Why the core-shell models should have two spring constants k1 and k2?

- Only the parameters for Li-O, Ni-O and O-O are given, do you have mixing rules for the others or how can they be set?

- What is the meaning of the number of fits? Is this meant for running the same procedure on a different set of structures? Or with different random seeds?

- In Training Set and Fitting, the directory is called 'vasprun' but should be 'vaspruns'. Also in the 'random_set_of_structures' function, if the optional parameter 'seed' is omitted then the function returns an empty list. Later, in the section 'Randomly selecting sets of structures', it is said that the default is None, while it is False.

Moreover, there are some details to be fixed:
- The name of the package is not consistent through the paper, the repository and the documentation. popoff, PopOff and POPOFF have been used. Please use the same name everywhere (also in the paths in the documentation for example).

- In the paper, give the meaning of the abbreviation NMC.

- Overall there are several orthographic mistakes (in the paper: 'fittted', in the readmes and the docs: 'potenial', 'relivant', 'ca be found in', in the notebook: 'paralising', 'agian', 'corespond', 'trianing', 'creat').

- The 'Module contents' section of the PopOff submodules in the docs is empty, and the link to 'Training set guide' for VASP doesn't work.

- The command to unzip the vasp files in the tests directory should be `gzip -d <file>`. It would also be interesting to briefly explain what is tested and how.

@dpsanders I think I didn't click on the acceptation link on time and cannot modify the above todo list, would it be possible to change this? Thank you
