### 2022-04-09 #3747 @arfon
:wave: folks, I'm taking over this submission on behalf of @dpsanders who is stepping down from the JOSS editorial team due to new work commitments. Thanks David for all of your amazing work as a JOSS editor ✨

@marshallmcdonnell, @lscalfi, @LMMorgan  – how are we getting along here? It looks like both @marshallmcdonnell and @lscalfi are somewhat through their reviews but haven't completed them yet.

@LMMorgan - there's lots of great feedback from @lscalfi in [this comment above](https://github.com/openjournals/joss-reviews/issues/3747#issuecomment-1013932948) – have you started working on addressing that feedback?
