### 2022-06-16 #3747 @arfon
@LMMorgan – it has been over six months since you've been active on this review. At this point it would appear that you are no-longer interested in pursuing a publication in JOSS.

If we don't hear back from you in the next two weeks, I will assume the submission is abandoned, and proceed to reject this submission.
